requirements.txt
discord.py
mysql-connector-python~=8.0.32
requests
dpy-http-server